backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Social Media Scrapers
praw==7.7.1
snscrape==0.7.0.20230622
PyGithub==1.59.1
instaloader==4.10.3
newspaper3k==0.2.8

# AI/ML Models
transformers==4.36.0
torch==2.1.0
spacy==3.7.2
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
wordcloud==1.9.2
networkx==3.2.1

# Database
supabase==2.0.2
prisma==0.11.0

# Report Generation
weasyprint==60.2
jinja2==3.1.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
